requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cymysql==0.9.14
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-appconf==1.0.3
django-axes==5.0.19
django-cors-headers==3.2.0
django-cymysql==2.2.0
django-extensions==2.2.1
django-filter==2.2.0
django-ipware==2.1.0
django-pagination==1.0.7
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-smtp-ssl==1.0
django-sslserver==0.20
djangorestframework==3.10.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.20
protobuf==3.12.2
oauthlib==3.1.0
openapi-codec==1.3.2
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
Werkzeug==0.15.5
wrapt==1.11.2